resume_parser/challenge-fernando/requirements.txt
fastapi==0.109.2
python-multipart==0.0.9
uvicorn==0.27.1
python-magic==0.4.27
google-generativeai==0.3.2
Pillow==10.2.0
pdf2image==1.17.0
boto3==1.34.34
python-jose==3.3.0
pytest==8.0.2
pytest-cov==4.1.0
httpx==0.26.0 